poc implement inspect opcode
to demonstrate https://mail.openvswitch.org/pipermail/ovs-discuss/2023-July/052591.html

we can do some minimal packet inspection so to allow or drop a
connection based on the SNI that is sent in the payload of the TLS
handshake. For that we identify the first packet, with the help of
conntracks: we mark a new connection with the 0x10 bit. The first TCP
PSH packet of an egress connection having the 0x10 mark is the packet we
want, and we clear that bit then.

Then we implement an inspect() action so that this packet is given to
a controller for "Next-generation" analysis, and doing a "pause" meanwhile.

the controller inspect this payload
Here northd adds inspect(glob="example.com") on each egress ACL.

We implement another controller in rust for this purpose
(for example this one
https://github.com/freedge/rust_ofp)